requirements.txt
opencv-python
ffmpeg-python
scipy
tensorboard 
requests
pillow
sklearn
wandb
matplotlib
timm
numpy
prettytable
torchsummary
torchinfo
seaborn
pandas
flow-vis
sailenv
tdw==1.9.1.0
scikit-image
gdown